# Performance: Comprehensive tooling optimizations for local dev and CI