--- a/.github/workflows/linting-and-testing.yml
+++ b/.github/workflows/linting-and-testing.yml
@@ -7,7 +7,6 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-
   lint:
     name: 🔬 Lint
     runs-on: ubuntu-latest
@@ -29,16 +28,26 @@ jobs:
           node-version: 22
           cache: "pnpm"
 
+      - name: 📦 Cache pnpm store
+        uses: actions/cache@v4
+        with:
+          path: ~/.pnpm-store
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            ${{ runner.os }}-pnpm-store-
+
       - name: 📦 Cache ESLint
         uses: actions/cache@v4
         with:
-          path: .eslintcache
+          path: |
+            .turbo/.eslintcache-ui
+            .turbo/.eslintcache-eslint-config
           key: ${{ runner.os }}-eslint-${{ hashFiles('**/package.json', '**/.eslintrc*', '**/eslint.config.*') }}
           restore-keys: |
             ${{ runner.os }}-eslint-
 
       - name: 📥 Install dependencies
-        run: pnpm install
+        run: pnpm install --frozen-lockfile
 
       - name: 🚀 Setup Turbo Cache
         uses: actions/cache@v4
@@ -88,6 +97,14 @@ jobs:
           node-version: 22
           cache: "pnpm"
 
+      - name: 📦 Cache pnpm store
+        uses: actions/cache@v4
+        with:
+          path: ~/.pnpm-store
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            ${{ runner.os }}-pnpm-store-
+
       - name: 🔧 Cache TypeScript build info
         uses: actions/cache@v4
         with:
@@ -99,7 +116,7 @@ jobs:
             ${{ runner.os }}-tsbuildinfo-
 
       - name: 📥 Install dependencies
-        run: pnpm install
+        run: pnpm install --frozen-lockfile
 
       - name: 🚀 Setup Turbo Cache
         uses: actions/cache@v4
@@ -149,6 +166,14 @@ jobs:
           node-version: 22
           cache: "pnpm"
 
+      - name: 📦 Cache pnpm store
+        uses: actions/cache@v4
+        with:
+          path: ~/.pnpm-store
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            ${{ runner.os }}-pnpm-store-
+
       - name: 🗄️ Cache Vite build outputs
         uses: actions/cache@v4
         with:
@@ -171,7 +196,7 @@ jobs:
             ${{ runner.os }}-vitest-
 
       - name: 📥 Install dependencies
-        run: pnpm install
+        run: pnpm install --frozen-lockfile
 
       - name: 🚀 Setup Turbo Cache
         uses: actions/cache@v4