--- a/.github/workflows/linting-and-testing.yml
+++ b/.github/workflows/linting-and-testing.yml
@@ -7,34 +7,68 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  lint:
-    name: 🔬 Lint
+  # First job: Build everything once
+  build:
+    name: 🏗️ Build
     runs-on: ubuntu-latest
     steps:
       - name: ⏱️ Start Timer
         run: echo "JOB_START=$(date +%s)" >> $GITHUB_ENV
 
-      - name: ⬇️ Checkout repo
-        uses: actions/checkout@v4
+      - name: 🔧 Setup Monorepo
+        uses: ./.github/actions/setup-monorepo
 
-      - uses: pnpm/action-setup@v4
-        name: Install pnpm
-        with:
-          run_install: false
-
-      - name: 📦 Install Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: "pnpm"
-
-      - name: 📦 Cache pnpm store
+      - name: 🔧 Cache TypeScript build info
         uses: actions/cache@v4
         with:
-          path: ~/.pnpm-store
-          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          path: |
+            **/*.tsbuildinfo
+            **/tsconfig.tsbuildinfo
+            packages/*/dist
+          key: ${{ runner.os }}-build-${{ github.sha }}
           restore-keys: |
-            ${{ runner.os }}-pnpm-store-
+            ${{ runner.os }}-build-
+
+      - name: 🏗️ Build all packages
+        id: build
+        run: |
+          echo "## 🏗️ Build Results" >> $GITHUB_STEP_SUMMARY
+          if pnpm run build; then
+            echo "✅ **Build completed successfully**" >> $GITHUB_STEP_SUMMARY
+            echo "status=success" >> $GITHUB_OUTPUT
+          else
+            echo "❌ **Build failed** - Check the logs for errors" >> $GITHUB_STEP_SUMMARY
+            echo "status=failure" >> $GITHUB_OUTPUT
+            exit 1
+          fi
+
+          # Calculate runtime
+          END_TIME=$(date +%s)
+          DURATION=$((END_TIME - JOB_START))
+          echo "" >> $GITHUB_STEP_SUMMARY
+          echo "_Duration: ${DURATION}s_" >> $GITHUB_STEP_SUMMARY
+
+  lint:
+    name: 🔬 Lint
+    needs: build # Depend on build to reuse compiled packages
+    runs-on: ubuntu-latest
+    steps:
+      - name: ⏱️ Start Timer
+        run: echo "JOB_START=$(date +%s)" >> $GITHUB_ENV
+
+      - name: 🔧 Setup Monorepo
+        uses: ./.github/actions/setup-monorepo
+
+      - name: 📦 Restore build artifacts
+        uses: actions/cache@v4
+        with:
+          path: |
+            **/*.tsbuildinfo
+            **/tsconfig.tsbuildinfo
+            packages/*/dist
+          key: ${{ runner.os }}-build-${{ github.sha }}
+          restore-keys: |
+            ${{ runner.os }}-build-
 
       - name: 📦 Cache ESLint
         uses: actions/cache@v4
@@ -45,17 +79,6 @@ jobs:
           key: ${{ runner.os }}-eslint-${{ hashFiles('**/package.json', '**/.eslintrc*', '**/eslint.config.*') }}
           restore-keys: |
             ${{ runner.os }}-eslint-
-
-      - name: 📥 Install dependencies
-        run: pnpm install --frozen-lockfile
-
-      - name: 🚀 Setup Turbo Cache
-        uses: actions/cache@v4
-        with:
-          path: .turbo
-          key: ${{ runner.os }}-turbo-lint-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-turbo-lint-
 
       - name: 🔬 Run ESLint
         id: lint
@@ -78,59 +101,32 @@ jobs:
 
   typecheck:
     name: 🔎 Type Check
+    needs: build # Depend on build to reuse compiled packages
     runs-on: ubuntu-latest
     steps:
       - name: ⏱️ Start Timer
         run: echo "JOB_START=$(date +%s)" >> $GITHUB_ENV
 
-      - name: ⬇️ Checkout repo
-        uses: actions/checkout@v4
+      - name: 🔧 Setup Monorepo
+        uses: ./.github/actions/setup-monorepo
 
-      - uses: pnpm/action-setup@v4
-        name: Install pnpm
-        with:
-          run_install: false
-
-      - name: 📦 Install Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: "pnpm"
-
-      - name: 📦 Cache pnpm store
-        uses: actions/cache@v4
-        with:
-          path: ~/.pnpm-store
-          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
-          restore-keys: |
-            ${{ runner.os }}-pnpm-store-
-
-      - name: 🔧 Cache TypeScript build info
+      - name: 📦 Restore build artifacts
         uses: actions/cache@v4
         with:
           path: |
             **/*.tsbuildinfo
             **/tsconfig.tsbuildinfo
-          key: ${{ runner.os }}-tsbuildinfo-${{ hashFiles('**/tsconfig.json', '**/tsconfig.*.json') }}
+            packages/*/dist
+          key: ${{ runner.os }}-build-${{ github.sha }}
           restore-keys: |
-            ${{ runner.os }}-tsbuildinfo-
-
-      - name: 📥 Install dependencies
-        run: pnpm install --frozen-lockfile
-
-      - name: 🚀 Setup Turbo Cache
-        uses: actions/cache@v4
-        with:
-          path: .turbo
-          key: ${{ runner.os }}-turbo-build-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-turbo-build-
+            ${{ runner.os }}-build-
 
       - name: 🔎 Run TypeScript Check
         id: typecheck
         run: |
           echo "## 🔎 TypeScript Results" >> $GITHUB_STEP_SUMMARY
-          if pnpm run build; then
+          # Since we already built in the build job, just run tsc --noEmit for type checking
+          if pnpm -r exec tsc --noEmit; then
             echo "✅ **Type checking passed successfully**" >> $GITHUB_STEP_SUMMARY
             echo "status=success" >> $GITHUB_OUTPUT
           else
@@ -147,39 +143,32 @@ jobs:
 
   test:
     name: 🧪 Tests
+    needs: build # Tests might need built packages
     runs-on: ubuntu-latest
     steps:
       - name: ⏱️ Start Timer
         run: echo "JOB_START=$(date +%s)" >> $GITHUB_ENV
 
-      - name: ⬇️ Checkout repo
-        uses: actions/checkout@v4
+      - name: 🔧 Setup Monorepo
+        uses: ./.github/actions/setup-monorepo
 
-      - uses: pnpm/action-setup@v4
-        name: Install pnpm
-        with:
-          run_install: false
-
-      - name: 📦 Install Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: "pnpm"
-
-      - name: 📦 Cache pnpm store
+      - name: 📦 Restore build artifacts
         uses: actions/cache@v4
         with:
-          path: ~/.pnpm-store
-          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          path: |
+            **/*.tsbuildinfo
+            **/tsconfig.tsbuildinfo
+            packages/*/dist
+          key: ${{ runner.os }}-build-${{ github.sha }}
           restore-keys: |
-            ${{ runner.os }}-pnpm-store-
+            ${{ runner.os }}-build-
 
       - name: 🗄️ Cache Vite build outputs
         uses: actions/cache@v4
         with:
           path: |
-            dist
-            .vite
+            packages/*/dist
+            packages/*/.vite
           key: ${{ runner.os }}-vite-${{ github.sha }}
           restore-keys: |
             ${{ runner.os }}-vite-
@@ -188,23 +177,12 @@ jobs:
         uses: actions/cache@v4
         with:
           path: |
-            node_modules/.vitest
-            coverage
+            packages/*/.vitest
+            packages/*/coverage
             test-results
           key: ${{ runner.os }}-vitest-${{ hashFiles('**/src/**/*.test.*', '**/src/**/*.spec.*', '**/vitest.config.*') }}
           restore-keys: |
             ${{ runner.os }}-vitest-
-
-      - name: 📥 Install dependencies
-        run: pnpm install --frozen-lockfile
-
-      - name: 🚀 Setup Turbo Cache
-        uses: actions/cache@v4
-        with:
-          path: .turbo
-          key: ${{ runner.os }}-turbo-test-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-turbo-test-
 
       - name: 🧪 Run Unit Tests
         id: tests
@@ -236,16 +214,16 @@ jobs:
   summary:
     name: 📊 Summary
     runs-on: ubuntu-latest
-    needs: [lint, typecheck, test]
+    needs: [build, lint, typecheck, test]
     if: always()
     steps:
       - name: 📊 Generate Summary Report
         run: |
-          echo "## 📊 Parallel Checks Summary" >> $GITHUB_STEP_SUMMARY
+          echo "## 📊 CI Pipeline Summary" >> $GITHUB_STEP_SUMMARY
           echo "" >> $GITHUB_STEP_SUMMARY
 
           # Determine overall status
-          if [[ "${{ needs.lint.result }}" == "success" && "${{ needs.typecheck.result }}" == "success" && "${{ needs.test.result }}" == "success" ]]; then
+          if [[ "${{ needs.build.result }}" == "success" && "${{ needs.lint.result }}" == "success" && "${{ needs.typecheck.result }}" == "success" && "${{ needs.test.result }}" == "success" ]]; then
             echo "### ✅ All checks passed!" >> $GITHUB_STEP_SUMMARY
             OVERALL_STATUS="success"
           else
@@ -256,6 +234,15 @@ jobs:
 
           echo "| Check | Status | Details |" >> $GITHUB_STEP_SUMMARY
           echo "|-------|--------|---------|" >> $GITHUB_STEP_SUMMARY
+
+          # Build status
+          if [[ "${{ needs.build.result }}" == "success" ]]; then
+            echo "| 🏗️ Build | ✅ Passed | All packages built successfully |" >> $GITHUB_STEP_SUMMARY
+          elif [[ "${{ needs.build.result }}" == "failure" ]]; then
+            echo "| 🏗️ Build | ❌ Failed | See [build job logs](${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}) |" >> $GITHUB_STEP_SUMMARY
+          else
+            echo "| 🏗️ Build | ⏭️ Skipped | - |" >> $GITHUB_STEP_SUMMARY
+          fi
 
           # Lint status
           if [[ "${{ needs.lint.result }}" == "success" ]]; then
@@ -285,10 +272,11 @@ jobs:
           fi
 
           echo "" >> $GITHUB_STEP_SUMMARY
-          echo "### ⚡ Performance Benefits" >> $GITHUB_STEP_SUMMARY
-          echo "- **Parallel Execution**: All checks run simultaneously" >> $GITHUB_STEP_SUMMARY
-          echo "- **Fail-Fast Disabled**: All checks complete even if one fails" >> $GITHUB_STEP_SUMMARY
-          echo "- **Optimized Caching**: Each job uses specific caches for better performance" >> $GITHUB_STEP_SUMMARY
+          echo "### ⚡ Performance Optimizations" >> $GITHUB_STEP_SUMMARY
+          echo "- **Shared Build Artifacts**: Build once, reuse across jobs" >> $GITHUB_STEP_SUMMARY
+          echo "- **Composite Actions**: Reduced setup duplication" >> $GITHUB_STEP_SUMMARY
+          echo "- **Smart Caching**: pnpm store, ESLint, TypeScript, and Turbo caches" >> $GITHUB_STEP_SUMMARY
+          echo "- **Parallel Execution**: Lint, typecheck, and test run in parallel after build" >> $GITHUB_STEP_SUMMARY
           echo "" >> $GITHUB_STEP_SUMMARY
 
           echo "---" >> $GITHUB_STEP_SUMMARY