--- a/.github/workflows/linting-and-testing.yml
+++ b/.github/workflows/linting-and-testing.yml
@@ -7,50 +7,8 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  # First job: Build everything once
-  build:
-    name: 🏗️ Build
-    runs-on: ubuntu-latest
-    steps:
-      - name: ⏱️ Start Timer
-        run: echo "JOB_START=$(date +%s)" >> $GITHUB_ENV
-
-      - name: 🔧 Setup Monorepo
-        uses: ./.github/actions/setup-monorepo
-
-      - name: 🔧 Cache TypeScript build info
-        uses: actions/cache@v4
-        with:
-          path: |
-            **/*.tsbuildinfo
-            **/tsconfig.tsbuildinfo
-            packages/*/dist
-          key: ${{ runner.os }}-build-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-build-
-
-      - name: 🏗️ Build all packages
-        id: build
-        run: |
-          echo "## 🏗️ Build Results" >> $GITHUB_STEP_SUMMARY
-          if pnpm run build; then
-            echo "✅ **Build completed successfully**" >> $GITHUB_STEP_SUMMARY
-            echo "status=success" >> $GITHUB_OUTPUT
-          else
-            echo "❌ **Build failed** - Check the logs for errors" >> $GITHUB_STEP_SUMMARY
-            echo "status=failure" >> $GITHUB_OUTPUT
-            exit 1
-          fi
-
-          # Calculate runtime
-          END_TIME=$(date +%s)
-          DURATION=$((END_TIME - JOB_START))
-          echo "" >> $GITHUB_STEP_SUMMARY
-          echo "_Duration: ${DURATION}s_" >> $GITHUB_STEP_SUMMARY
-
   lint:
     name: 🔬 Lint
-    needs: build # Depend on build to reuse compiled packages
     runs-on: ubuntu-latest
     steps:
       - name: ⏱️ Start Timer
@@ -58,17 +16,6 @@ jobs:
 
       - name: 🔧 Setup Monorepo
         uses: ./.github/actions/setup-monorepo
-
-      - name: 📦 Restore build artifacts
-        uses: actions/cache@v4
-        with:
-          path: |
-            **/*.tsbuildinfo
-            **/tsconfig.tsbuildinfo
-            packages/*/dist
-          key: ${{ runner.os }}-build-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-build-
 
       - name: 📦 Cache ESLint
         uses: actions/cache@v4
@@ -79,6 +26,9 @@ jobs:
           key: ${{ runner.os }}-eslint-${{ hashFiles('**/package.json', '**/.eslintrc*', '**/eslint.config.*') }}
           restore-keys: |
             ${{ runner.os }}-eslint-
+
+      - name: 🏗️ Build eslint-config package
+        run: pnpm --filter @protomolecule/eslint-config build
 
       - name: 🔬 Run ESLint
         id: lint
@@ -101,7 +51,6 @@ jobs:
 
   typecheck:
     name: 🔎 Type Check
-    needs: build # Depend on build to reuse compiled packages
     runs-on: ubuntu-latest
     steps:
       - name: ⏱️ Start Timer
@@ -110,23 +59,22 @@ jobs:
       - name: 🔧 Setup Monorepo
         uses: ./.github/actions/setup-monorepo
 
-      - name: 📦 Restore build artifacts
+      - name: 🔧 Cache TypeScript build info
         uses: actions/cache@v4
         with:
           path: |
             **/*.tsbuildinfo
             **/tsconfig.tsbuildinfo
-            packages/*/dist
-          key: ${{ runner.os }}-build-${{ github.sha }}
+          key: ${{ runner.os }}-tsbuildinfo-${{ hashFiles('**/tsconfig.json', '**/tsconfig.*.json') }}
           restore-keys: |
-            ${{ runner.os }}-build-
+            ${{ runner.os }}-tsbuildinfo-
 
       - name: 🔎 Run TypeScript Check
         id: typecheck
         run: |
           echo "## 🔎 TypeScript Results" >> $GITHUB_STEP_SUMMARY
-          # Since we already built in the build job, just run tsc --noEmit for type checking
-          if pnpm -r exec tsc --noEmit; then
+          # Run the build command which includes type checking
+          if pnpm run build; then
             echo "✅ **Type checking passed successfully**" >> $GITHUB_STEP_SUMMARY
             echo "status=success" >> $GITHUB_OUTPUT
           else
@@ -143,7 +91,6 @@ jobs:
 
   test:
     name: 🧪 Tests
-    needs: build # Tests might need built packages
     runs-on: ubuntu-latest
     steps:
       - name: ⏱️ Start Timer
@@ -151,27 +98,6 @@ jobs:
 
       - name: 🔧 Setup Monorepo
         uses: ./.github/actions/setup-monorepo
-
-      - name: 📦 Restore build artifacts
-        uses: actions/cache@v4
-        with:
-          path: |
-            **/*.tsbuildinfo
-            **/tsconfig.tsbuildinfo
-            packages/*/dist
-          key: ${{ runner.os }}-build-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-build-
-
-      - name: 🗄️ Cache Vite build outputs
-        uses: actions/cache@v4
-        with:
-          path: |
-            packages/*/dist
-            packages/*/.vite
-          key: ${{ runner.os }}-vite-${{ github.sha }}
-          restore-keys: |
-            ${{ runner.os }}-vite-
 
       - name: 🧪 Cache Test Results
         uses: actions/cache@v4
@@ -214,7 +140,7 @@ jobs:
   summary:
     name: 📊 Summary
     runs-on: ubuntu-latest
-    needs: [build, lint, typecheck, test]
+    needs: [lint, typecheck, test]
     if: always()
     steps:
       - name: 📊 Generate Summary Report
@@ -223,7 +149,7 @@ jobs:
           echo "" >> $GITHUB_STEP_SUMMARY
 
           # Determine overall status
-          if [[ "${{ needs.build.result }}" == "success" && "${{ needs.lint.result }}" == "success" && "${{ needs.typecheck.result }}" == "success" && "${{ needs.test.result }}" == "success" ]]; then
+          if [[ "${{ needs.lint.result }}" == "success" && "${{ needs.typecheck.result }}" == "success" && "${{ needs.test.result }}" == "success" ]]; then
             echo "### ✅ All checks passed!" >> $GITHUB_STEP_SUMMARY
             OVERALL_STATUS="success"
           else
@@ -234,15 +160,6 @@ jobs:
 
           echo "| Check | Status | Details |" >> $GITHUB_STEP_SUMMARY
           echo "|-------|--------|---------|" >> $GITHUB_STEP_SUMMARY
-
-          # Build status
-          if [[ "${{ needs.build.result }}" == "success" ]]; then
-            echo "| 🏗️ Build | ✅ Passed | All packages built successfully |" >> $GITHUB_STEP_SUMMARY
-          elif [[ "${{ needs.build.result }}" == "failure" ]]; then
-            echo "| 🏗️ Build | ❌ Failed | See [build job logs](${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}) |" >> $GITHUB_STEP_SUMMARY
-          else
-            echo "| 🏗️ Build | ⏭️ Skipped | - |" >> $GITHUB_STEP_SUMMARY
-          fi
 
           # Lint status
           if [[ "${{ needs.lint.result }}" == "success" ]]; then
@@ -272,11 +189,11 @@ jobs:
           fi
 
           echo "" >> $GITHUB_STEP_SUMMARY
-          echo "### ⚡ Performance Optimizations" >> $GITHUB_STEP_SUMMARY
-          echo "- **Shared Build Artifacts**: Build once, reuse across jobs" >> $GITHUB_STEP_SUMMARY
+          echo "### ⚡ Performance Features" >> $GITHUB_STEP_SUMMARY
+          echo "- **True Parallel Execution**: All checks run simultaneously from start" >> $GITHUB_STEP_SUMMARY
           echo "- **Composite Actions**: Reduced setup duplication" >> $GITHUB_STEP_SUMMARY
           echo "- **Smart Caching**: pnpm store, ESLint, TypeScript, and Turbo caches" >> $GITHUB_STEP_SUMMARY
-          echo "- **Parallel Execution**: Lint, typecheck, and test run in parallel after build" >> $GITHUB_STEP_SUMMARY
+          echo "- **Minimal Building**: Only builds what's necessary (eslint-config for linting)" >> $GITHUB_STEP_SUMMARY
           echo "" >> $GITHUB_STEP_SUMMARY
 
           echo "---" >> $GITHUB_STEP_SUMMARY