--- a/.gitignore
+++ b/.gitignore
@@ -32,3 +32,7 @@ notes/*
 
 # Turborepo
 .turbo
+
+# TypeScript build info
+*.tsbuildinfo
+.tsbuildinfo