--- a/package.json
+++ b/package.json
@@ -21,9 +21,19 @@
   "devDependencies": {
     "@types/node": "^20.12.12",
     "husky": "^9.1.7",
+    "lint-staged": "^16.1.6",
     "rimraf": "^6.0.1",
     "turbo": "^2.5.6",
     "vitest": "^3.2.4"
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx,mjs}": [
+      "eslint --fix --cache",
+      "prettier --write"
+    ],
+    "*.{json,md,yml,yaml,css}": [
+      "prettier --write"
+    ]
   },
   "packageManager": "pnpm@10.15.0",
   "engines": {