--- a/packages/ui/.storybook/main.ts
+++ b/packages/ui/.storybook/main.ts
@@ -3,7 +3,6 @@ import { type StorybookConfig } from "@storybook/react-vite";
 
 const config: StorybookConfig = {
   addons: [
-    "@storybook/addon-onboarding",
     "@storybook/addon-links",
     "@chromatic-com/storybook",
     "@storybook/addon-docs",