--- a/packages/ui/package.json
+++ b/packages/ui/package.json
@@ -49,7 +49,6 @@
     "@protomolecule/tsconfig": "workspace:*",
     "@storybook/addon-docs": "^9.1.3",
     "@storybook/addon-links": "^9.1.3",
-    "@storybook/addon-onboarding": "^9.1.3",
     "@storybook/react-vite": "^9.1.3",
     "@tailwindcss/vite": "^4.1.12",
     "@testing-library/dom": "^10.1.0",